Changed services to be added by path & loaded as modules
